!!CRITICAL!! Fix Power Star Mask "Memory Leak"
There's no actual Memory Leak.
Collecting a Masked Power Star had the ability to change the name of the galaxy that you were in, which would let you collect the wrong star from a galaxy. Almost like a wrong warp, but for getting stars. This is fixed by simply allowing the Power Star Mask code to reset the galaxy name instead of only changing it.