Fixed a Long Standing bug with StageInfo
This has been a problem presumably since GLE-V2 where the game would never even try to check the StageInfo overrides of the zone that you currently had a checkpoint in because the wrong address was assumed to be the checkpoint data. This has now been fixed, so StageInfo overrides work as intended now.